+++ Inc/Utils/Random.hpp
@@ -9,7 +9,7 @@ namespace Galactic
     {
         float Rand(float min, float max)
         {
-            static std::mt19937 m_gen((size_t)std::chrono::system_clock::now().time_since_epoch().count());
+            static std::mt19937 m_gen((unsigned int)std::chrono::system_clock::now().time_since_epoch().count());
 
             std::uniform_real_distribution<double> dist(min, max);
